# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: СМП (Система медленных платежей) API
  version: 1.0.0
  description: |
    Единая спецификация API-контрактов для взаимодействия микросервисов СМП.
servers:
  - url: http://localhost:8080
    description: Gateway (единственная точка входа для внешних клиентов)
  - url: http://localhost:8081
    description: Switch Service
  - url: http://localhost:8082
    description: Authorization Service
  - url: http://localhost:8083
    description: Card Management Service
  - url: http://localhost:8084
    description: Terminal Simulator
  - url: http://localhost:8085
    description: Merchant Simulator
  - url: http://localhost:8088
    description: Transaction Logger
...
paths:
...
  /api/transactions:
    post:
      tags: [Transactions]
      summary: Отправить запрос на авторизацию транзакции
      description:
        Главный входной эндпоинт. Gateway валидирует запрос и отправляет его в Switch.
        Switch определяет эмитента по BIN, передаёт запрос в Authorization Service,
        получает решение и записывает транзакцию в Logger. Возвращает итоговый ответ авторизации.
      operationId: authorizeTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationRequest'
      responses:
        '200':
          description: Результат авторизации (успех или отказ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationResponse'
        '400':
          description: Ошибка валидации запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "VALIDATION_ERROR"
                message: "Field 'pan' must be exactly 16 digits"
                timestamp: "2026-06-01T10:30:00Z"
        '429':
          description: Превышен rate-limit (100 запросов в секунду)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "RATE_LIMIT_EXCEEDED"
                message: "Too many requests. Try again later."
                retryAfterMs: 1000
        '503':
          description: Downstream-сервис недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "SERVICE_UNAVAILABLE"
                message: "Authorization service is temporarily unavailable"
                serviceName: "authorization"
  /api/transactions/search:
    get:
      tags: [Transactions]
      summary: Поиск транзакций с фильтрацией и постраничной навигацией
      description: |
        Поиск по журналу транзакций. Все фильтры опциональны и объединяются через AND.
        Пагинация offset-based; `limit` (размер страницы) и `offset` (смещение от начала).
      operationId: searchTransactions
      parameters:
        - name: pan
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{16}$'
          description: Фильтр по номеру карты
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [APPROVED, DECLINED]
          description: Фильтр по итоговому статусу транзакции
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: |
            Нижняя граница по дате транзакции (включительно).
            Пример: `2026-06-01`
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: |
            Верхняя граница по дате транзакции (включительно, до конца суток).
            Пример: `2026-06-01`
        - name: merchantId
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по ID мерчанта (торговой точки)
        - name: issuerId
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по ID банка-эмитента
        - name: mcc
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}$'
          description: Фильтр по коду категории мерчанта (MCC)
        - name: declineReason
          in: query
          required: false
          schema:
            type: string
          description: |
            Полнотекстовый поиск по причине отказа.
            Перед поиском `_` и `-` нормализуются в пробелы, регистр игнорируется.
            Примеры совпадающих запросов для `CARD_NOT_FOUND`:
            `CARD_NOT_FOUND`, `card_not_found`, `card not found`, `card`, `found`.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
          description: Количество транзакций на странице
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Смещение от начала выборки (offset-пагинация)
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionSearchResponse'
        '400':
          description: Невалидные параметры (например, `limit <= 0` или `offset < 0`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/transactions/export:
    get:
      tags: [ Transactions ]
      summary: Экспорт транзакций в CSV
      description: |
        Выгружает транзакции, удовлетворяющие фильтру, в CSV (сортировка createdAt DESC).
        Параметры фильтра те же, что у /api/transactions/search, кроме limit/offset — они не применяются.
        Экспорт ограничен 50 000 строк: если под фильтр попадает больше, выгружаются первые 50 000.
      operationId: exportTransactions
      parameters:
        - name: pan
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{16}$'
          description: Фильтр по номеру карты
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ APPROVED, DECLINED ]
          description: Фильтр по итоговому статусу транзакции
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Нижняя граница по дате транзакции (включительно)
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Верхняя граница по дате транзакции (включительно, до конца суток)
        - name: merchantId
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по ID мерчанта
        - name: issuerId
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по ID банка-эмитента
        - name: mcc
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}$'
          description: Фильтр по коду категории мерчанта (MCC)
        - name: declineReason
          in: query
          required: false
          schema:
            type: string
          description: Полнотекстовый поиск по причине отказа
      responses:
        '200':
          description: CSV-файл с транзакциями
          content:
            text/csv:
              schema:
                type: string
        '400':
          description: Невалидные параметры фильтра
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error, message, timestamp]
      properties:
        error:
          type: string
          description: Машинно-читаемый код ошибки
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Человеко-читаемое описание ошибки
          example: "Field 'pan' must be exactly 16 digits"
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
        serviceName:
          type: string
          description: Имя недоступного сервиса
        retryAfterMs:
          type: integer
          description: Через сколько миллисекунд повторить запрос
...
    AuthorizationRequest:
      type: object
      required:
        - mti
        - stan
        - pan
        - processingCode
        - amount
        - currencyCode
        - transmissionDateTime
        - terminalId
        - merchantId
        - mcc
        - acquirerId
      properties:
        mti:
          type: string
          pattern: '^\d{4}$'
          description: |
            Message Type Identifier - тип сообщения ISO 8583
          example: "0100"
        stan:
          type: string
          pattern: '^\d{6}$'
          description: System Trace Audit Number - уникальный номер транзакции в терминале (000001..999999)
          example: "000001"
        pan:
          type: string
          pattern: '^\d{16}$'
          description: Primary Account Number - номер карты (16 цифр)
          example: "4000001234560001"
        processingCode:
          type: string
          pattern: '^\d{6}$'
          description: Код обработки (000000 = покупка, 000001 = снятие наличных)
          example: "000000"
        amount:
          type: integer
          minimum: 1
          description: Сумма транзакции в минорных единицах (копейках). 150000 = 1 500.00 ₽
          example: 150000
        currencyCode:
          type: string
          pattern: '^\d{3}$'
          description: Код валюты ISO 4217 (643 = RUB)
          example: "643"
        transmissionDateTime:
          type: string
          format: date-time
          description: Дата и время отправки транзакции терминалом (ISO 8601, UTC)
        terminalId:
          type: string
          maxLength: 8
          description: Идентификатор POS-терминала
          example: "TERM001"
        terminalType:
          type: string
          enum: [POS, ATM, ECOM]
          description: Тип терминала (опционально)
        merchantId:
          type: string
          maxLength: 15
          description: Идентификатор мерчанта (торговой точки)
          example: "MERCH12345678901"
        mcc:
          type: string
          pattern: '^\d{4}$'
          description: Merchant Category Code - код категории бизнеса
          example: "5411"
        acquirerId:
          type: string
          description: ID банка-эквайера
          example: "ACQ001"
        issuerId:
          type: string
          description: |
            ID банка-эмитента. Не заполняется клиентом - определяется Switch-ом
            по BIN карты на этапе маршрутизации.
          example: "ISS001"
    AuthorizationResponse:
      type: object
      required:
        - mti
        - stan
        - rrn
        - authCode
        - responseCode
        - status
      properties:
        mti:
          type: string
          example: "0110"
          description: Message Type Identifier ответа (0110 = ответ на авторизацию 0100)
        stan:
          type: string
          description: STAN из исходного запроса (для сопоставления)
        rrn:
          type: string
          maxLength: 12
          description: |
            Retrieval Reference Number - уникальный идентификатор транзакции в системе.
            Формат: `{YDDD}{HH}{mm}{ss}{seq}`. Уникален в рамках всей системы.
          example: "012345678901"
        authCode:
          type: string
          maxLength: 6
          description: Код авторизации (6 символов A-Z0-9). Генерируется только при APPROVED.
          example: "ABC123"
        responseCode:
          type: string
          pattern: '^\d{2}$'
          description: |
            Код ответа ISO 8583:
            * `00` - Approved (одобрено)
            * `05` - Do Not Honor (отказано, общая причина)
            * `14` - Invalid Card Number (карта не найдена / неверный BIN)
            * `51` - Insufficient Funds (недостаточно средств)
            * `54` - Expired Card (карта просрочена)
            * `61` - Exceeds Amount Limit (превышен дневной/месячный лимит)
            * `96` - System Error (системная ошибка, например, откат из-за недоступности Logger)
          example: "00"
        status:
          type: string
          enum: [APPROVED, DECLINED]
          description: Итоговый статус транзакции
        declineReason:
          type: string
          nullable: true
          description: |
            Человеко-читаемая причина отклонения. Заполняется только при status=DECLINED.
            Примеры: "Insufficient funds", "CARD_BLOCKED", "ISSUER_TIMEOUT".
          example: "Insufficient funds"
        processingTimeMs:
          type: integer
          description: Время обработки в Authorization Service, мс.
          example: 42
...
    Transaction:
      type: object
      description: |
        Полная запись о транзакции в журнале. Объединяет AuthorizationRequest +
        AuthorizationResponse + системные поля (id, acquiringFee, createdAt).
      allOf:
        - $ref: '#/components/schemas/AuthorizationRequest'
        - $ref: '#/components/schemas/AuthorizationResponse'
        - type: object
          required: [id, createdAt]
          properties:
            id:
              type: string
              format: uuid
              description: Уникальный идентификатор записи транзакции в БД Logger
            acquiringFee:
              type: integer
              nullable: true
              description: Комиссия эквайера в копейках (рассчитывается Merchant Simulator)
            createdAt:
              type: string
              format: date-time
              description: Дата и время записи транзакции в Logger
    TransactionSearchResponse:
      type: object
      required: [total, transactions]
      properties:
        total:
          type: integer
          format: int64
          description: Общее число транзакций, удовлетворяющих фильтру (без учёта пагинации)
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...